Added register command. More implementation for rp echos.
