+++ src/utils.h
@@ -340,6 +340,8 @@ extern bool PLR_TOG_CHK(char_data *ch, dword offset);
 #define GET_WHOTITLE(ch) ((ch)->player.whotitle)
 #define GET_LEVEL(ch)   ((ch)->player.level)
 #define GET_PASSWD(ch)  ((ch)->player.passwd)
+#define GET_EMAIL(ch)   ((ch)->player.email ? (ch)->player.email : "not set")
+#define SETTABLE_EMAIL(ch)   ((ch)->player.email)
 
 /*
  * I wonder if this definition of GET_REAL_LEVEL should be the definition
